fixes from mac

+++ src/core/primitives.cc
@@ -229,7 +229,7 @@ std::unique_ptr<const Geometry> SphereNode::createGeometry() const
     polyset->indices.back().push_back(i);
   }
 
-  for (size_t i = 0; i < num_rings - 1; ++i) {
+  for (int i = 0; i < num_rings - 1; ++i) {
     for (int r=0;r<num_fragments;++r) {
       polyset->indices.push_back({
         i*num_fragments+(r+1)%num_fragments,
